+++ src/08_Config.ps1
@@ -315,31 +315,44 @@ function Invoke-PrintNightmareCheck {
     License: BSD 3-Clause
     
     .DESCRIPTION
-    Fully up-to-date machines are still vulnerable to the PrintNightmare exploit if the "Point and Print Restrictions" Group policy is configured to allow users to install printer drivers or add print servers without administrator privileges. More precisely, if "NoWarningNoElevationOnInstall" or "UpdatePromptSettings" is set to 1, the machine is vulnerable. There is one exception though. If the patch for CVE-2021-34527 was installed, the "RestrictDriverInstallationToAdministrators" parameter can be set to 1 (or a value greater than 1) to override the "Point and Print" settings. In this case, only administrators can install printer drivers or print servers, regardless of the two other values.
-    
-    .PARAMETER Info
-    Use this parameter to report any information about the "Point and Print" configuration, regardless of the vulnerability status.
+    Fully up-to-date machines are still vulnerable to the PrintNightmare exploit if the "Point and Print Restrictions" Group policy is configured to allow users to install printer drivers or add print servers without administrator privileges. More precisely, if "NoWarningNoElevationOnInstall" or "UpdatePromptSettings" is set to 1, the machine is vulnerable. There is one exception though. If the patch for CVE-2021-34527 was installed, the "RestrictDriverInstallationToAdministrators" parameter can be set to 1 (or a value greater than 1) to override the "Point and Print" settings. In this case, only administrators can install printer drivers or print servers, regardless of the two other values. The "PackagePointAndPrintOnly" setting can also be set to 1 to only allow drivers to be downloaded and installed from a predefined list of print servers.
     
     .EXAMPLE
     PS C:\> Invoke-PrintNightmareCheck
 
-    Path  : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
-    Value : NoWarningNoElevationOnInstall
-    Data  : 1
+    Key         : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
+    Value       : NoWarningNoElevationOnInstall
+    Data        : 1
+    Description : Do not show warning or elevation prompt
+    Compliance  : False
 
-    Path  : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
-    Value : RestrictDriverInstallationToAdministrators
-    Data  : 0
+    Key         : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
+    Value       : UpdatePromptSettings
+    Data        : 2
+    Description : Do not show warning or elevation prompt
+    Compliance  : False
+
+    Key         : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
+    Value       : RestrictDriverInstallationToAdministrators
+    Data        : 0
+    Description : Non-administrators can install print drivers
+    Compliance  : False
+
+    Key         : HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint
+    Value       : PackagePointAndPrintOnly
+    Data        : (null)
+    Description : Users are not restricted to package-aware point and print
+    Compliance  : False
 
     .LINK
     https://msrc.microsoft.com/update-guide/vulnerability/CVE-2021-34527
     https://support.microsoft.com/en-us/topic/kb5005010-restricting-installation-of-new-printer-drivers-after-applying-the-july-6-2021-updates-31b91c02-05bc-4ada-a7ea-183b129578a7
+    https://admx.help/?Category=Windows_10_2016&Policy=Microsoft.Policies.Printing::PointAndPrint_Restrictions_Win7
+    https://admx.help/?Category=PrintNightmare&Policy=PrintNightmare::POL_0F5609EA_BBB4_43FB_839A_231E44CEDD71
+    https://admx.help/?Category=Windows_10_2016&Policy=Microsoft.Policies.Printing::PackagePointAndPrintOnly
     #>
 
-    [CmdletBinding()] Param(
-        [switch]
-        $Info = $false
-    )
+    [CmdletBinding()] Param()
 
     # If the Print Spooler is not installed or is disabled, return immediately
     $Service = Get-ServiceList -FilterLevel 2 | Where-Object { $_.Name -eq "Spooler" }
@@ -349,60 +362,105 @@ function Invoke-PrintNightmareCheck {
     }
 
     $Results = @()
+
+    # If "NoWarningNoElevationOnInstall" is not set, the default value is 0, which means "Show warning
+    # and elevation prompt" (i.e. not vulnerable).
     $RegKey = "HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint"
-
-    # If "NoWarningNoElevationOnInstall" is not set, the default value is 0 (i.e. not vulnerable).
-    $Value = "NoWarningNoElevationOnInstall"
-    $ItemProptery = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $Value -ErrorAction SilentlyContinue -ErrorVariable GetItemProperty
-    if (-not $GetItemProperty) {
-        $WarningInstall = New-Object -TypeName PSObject
-        $WarningInstall | Add-Member -MemberType "NoteProperty" -Name "Path" -Value $RegKey
-        $WarningInstall | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $Value
-        $WarningInstall | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $ItemProptery.NoWarningNoElevationOnInstall
-        $WarningInstall | Add-Member -MemberType "NoteProperty" -Name "Vulnerable" -Value ($ItemProptery.NoWarningNoElevationOnInstall -ne 0)
-        [object[]]$Results += $WarningInstall
+    $RegValue = "NoWarningNoElevationOnInstall"
+    $RegItem = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $RegValue -ErrorAction SilentlyContinue
+    if ($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 0) {
+        $Description = "Show warning and elevation prompt"
     }
-
-    # If "UpdatePromptSettings" is not set, the default value is 0 (i.e. not vulnerable).
-    $Value = "UpdatePromptSettings"
-    $ItemProptery = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $Value -ErrorAction SilentlyContinue -ErrorVariable GetItemProperty
-    if (-not $GetItemProperty) {
-        $WarningUpdate = New-Object -TypeName PSObject
-        $WarningUpdate | Add-Member -MemberType "NoteProperty" -Name "Path" -Value $RegKey
-        $WarningUpdate | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $Value
-        $WarningUpdate | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $ItemProptery.UpdatePromptSettings
-        $WarningUpdate | Add-Member -MemberType "NoteProperty" -Name "Vulnerable" -Value ($ItemProptery.UpdatePromptSettings -ne 0)
-        [object[]]$Results += $WarningUpdate
+    else {
+        $Description = "Do not show warning or elevation prompt"
     }
+    Write-Verbose "NoWarningNoElevationOnInstall: $($Description)"
+    $Result = New-Object -TypeName PSObject
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Key" -Value $RegKey
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $RegValue
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $(if ($null -eq $RegItem.$RegValue) { "(null)" } else { $RegItem.$RegValue })
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Description" -Value $Description
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Compliance" -Value $($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 0)
+    [object[]]$Results += $Result
+
+    # If "UpdatePromptSettings" is not set, the default value is 0, which means "Show warning and 
+    # elevation prompt" (i.e. not vulnerable).
+    $RegKey = "HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint"
+    $RegValue = "UpdatePromptSettings"
+    $RegItem = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $RegValue -ErrorAction SilentlyContinue
+    if ($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 0) {
+        $Description = "Show warning and elevation prompt"
+    }
+    elseif ($RegItem.$RegValue -eq 1) {
+        $Description = "Show warning only"
+    }
+    elseif ($RegItem.$RegValue -eq 2) {
+        $Description = "Do not show warning or elevation prompt"
+    }
+    Write-Verbose "UpdatePromptSettings: $($Description)"
+    $Result = New-Object -TypeName PSObject
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Key" -Value $RegKey
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $RegValue
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $(if ($null -eq $RegItem.$RegValue) { "(null)" } else { $RegItem.$RegValue })
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Description" -Value $Description
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Compliance" -Value $($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 0)
+    [object[]]$Results += $Result
 
     # With the patch for CVE-2021-34527, MS added the "RestrictDriverInstallationToAdministrators"
     # setting. If this is set to 1 or any non-zero value then the Point and Print Restrictions Group
     # policy settings (i.e. the two previous registry values) are overridden and only administrators 
-    # can install printer drivers on a print server.
-    $Value = "RestrictDriverInstallationToAdministrators"
-    $ItemProptery = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $Value -ErrorAction SilentlyContinue -ErrorVariable GetItemProperty
-    if (-not $GetItemProperty) {
-        $RestrictInstall = New-Object -TypeName PSObject
-        $RestrictInstall | Add-Member -MemberType "NoteProperty" -Name "Path" -Value $RegKey
-        $RestrictInstall | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $Value
-        $RestrictInstall | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $ItemProptery.RestrictDriverInstallationToAdministrators
-        $RestrictInstall | Add-Member -MemberType "NoteProperty" -Name "Vulnerable" -Value ($ItemProptery.RestrictDriverInstallationToAdministrators -eq 0)
-        [object[]]$Results += $RestrictInstall
+    # can install printer drivers on a print server. "Updates released August 10, 2021 or later have 
+    # a default of 1 (enabled)."
+    $RegKey = "HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint"
+    $RegValue = "RestrictDriverInstallationToAdministrators"
+    $RegItem = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $RegValue -ErrorAction SilentlyContinue
+    if ($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 1) {
+        $Description = "Only administrators can install print drivers"
+    }
+    else {
+        $Description = "Non-administrators can install print drivers"
+    }
+    Write-Verbose "RestrictDriverInstallationToAdministrators: $($Description)"
+    $Result = New-Object -TypeName PSObject
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Key" -Value $RegKey
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $RegValue
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $(if ($null -eq $RegItem.$RegValue) { "(null)" } else { $RegItem.$RegValue })
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Description" -Value $Description
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Compliance" -Value $($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 1)
+    [object[]]$Results += $Result
+
+    # If "PackagePointAndPrintOnly" is enabled, clients can only install signed drivers from a 
+    # predefined list of print servers. This list is defined in thrhough the "ListofServers" subkey.
+    # The content of the regisry key should be checked manually.
+    $RegKey = "HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint"
+    $RegValue = "PackagePointAndPrintOnly"
+    $RegItem = Get-ItemProperty -Path "Registry::$($RegKey)" -Name $RegValue -ErrorAction SilentlyContinue
+    if ($null -eq $RegItem.$RegValue -or $RegItem.$RegValue -eq 0) {
+        $Description = "Users are not restricted to package-aware point and print only"
+    }
+    else {
+        $Description = "Users are restricted to package-aware point and print only (check 'HKLM\SOFTWARE\Policies\Microsoft\Windows NT\Printers\PointAndPrint\ListofServers')."
+    }
+    Write-Verbose "PackagePointAndPrintOnly: $($Description)"
+    $Result = New-Object -TypeName PSObject
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Key" -Value $RegKey
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Value" -Value $RegValue
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Data" -Value $(if ($null -eq $RegItem.$RegValue) { "(null)" } else { $RegItem.$RegValue })
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Description" -Value $Description
+    $Result | Add-Member -MemberType "NoteProperty" -Name "Compliance" -Value $($RegItem.$RegValue -eq 1)
+    [object[]]$Results += $Result
+    
+    # The system is vulnerable only if none of the checked items are compliant. So loop through all
+    # the results and return as soon as one of them is marked as compliant. Note: we could also
+    # immediately return after one item is seen as compliant, but this method allows us to have the
+    # whole picture, in case we were to transform this into an 'Info' check.
+    foreach ($Result in $Results) {
+        if ($Result.Compliance -eq $true) {
+            return
+        }
     }
 
-    if ($Info) { $Results; return }
-
-    # Not vulnerable if:
-    #  - RestrictDriverInstallationToAdministrators is set to 1
-    #  - NoWarningNoElevationOnInstall or UpdatePromptSettings are not set
-    #  - NoWarningNoElevationOnInstall is set to 0
-    #  - UpdatePromptSettings is set to 0
-    if ($RestrictInstall -and (-not $RestrictInstall.Vulnerable)) { return }
-    if (-not $WarningInstall) { return }
-    if (-not $WarningUpdate) { return }
-    if ((-not $WarningInstall.Vulnerable) -or (-not $WarningUpdate.Vulnerable)) { return }
-
-    $Results | Select-Object -Property Path,Value,Data
+    $Results | Select-Object -ExcludeProperty Compliance
 }
 
 function Invoke-DriverCoInstallersCheck {